move reflection stuff into its own folder; remove some hardcoded CurrentDomain stuff; add methods from #41
